refactor (TicketCount): 티켓 장수 0미만으로 내려가지 않도록 원시값 포장

+++ app/src/main/java/woowacourse/movie/movieTicket/MovieTicketActivity.kt
@@ -45,7 +45,7 @@ class MovieTicketActivity : AppCompatActivity() {
         with(ticket) {
             ticketTitleView.text = title
             ticketMovieReleaseDateView.text = DateUtil(context).getDate(date)
-            ticketCountView.text = getString(R.string.movie_ticket_count).format(count)
+            ticketCountView.text = getString(R.string.movie_ticket_count).format(count.toInt())
             ticketTotalPriceView.text = getString(R.string.movie_ticket_total_price).format(decimalFormat.format(getTotalPrice()))
         }
     }
